requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
aiogram>=3.3.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
greenlet>=3.0.0
google-genai>=1.0.0
openai>=1.68.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
jinja2>=3.1.3
requests>=2.31.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
alembic>=1.13.0
